build: ensure monorepo builds on vercel and render
