+++ Dockerfile
@@ -1,36 +1,31 @@
-# --- Base Stage ---
-FROM node:20-slim AS base
-WORKDIR /app
-COPY package*.json ./
-
-# Install only production dependencies by default
-RUN npm ci --only=production --legacy-peer-deps
-
-# --- Build Stage ---
+# Multi-stage build for monorepo (client + server)
 FROM node:20-slim AS build
 WORKDIR /app
+
+# Install dependencies including devDeps for both workspaces
+COPY package*.json ./
+COPY server/package*.json server/
+COPY client/package*.json client/
+RUN npm ci --legacy-peer-deps
+
+# Copy source and build
 COPY . .
-
-# Install ALL dependencies (including vite + devDeps) for build
-RUN npm install --legacy-peer-deps
-
-# Compile TypeScript for server and shared
 RUN npm run build
 
-# --- Production Stage ---
+# Production image
 FROM node:20-slim AS prod
 WORKDIR /app
 
-# Copy only built output & necessary files
+# Copy compiled output
 COPY --from=build /app/dist ./dist
-COPY package*.json ./
 
-# Install only production dependencies for runtime
+# Install only production dependencies for the server
+COPY server/package.json ./package.json
+COPY server/package-lock.json ./package-lock.json
 RUN npm ci --only=production --legacy-peer-deps
 
-# Use port from environment or default to 3000
-ENV PORT=${PORT:-3000}
-EXPOSE ${PORT}
+ENV PORT=3000
+EXPOSE 3000
 
-# Start the server
-CMD ["node", "dist/index.js"]
+CMD ["node", "dist/server/index.js"]
+
