+++ client/src/components/booking-form-modal.tsx
@@ -29,7 +29,7 @@ import {
   SelectValue,
 } from "@/components/ui/select";
 import { useToast } from "@/hooks/use-toast";
-import { apiRequest } from "@/lib/queryClient";
+import { apiRequest } from "@/lib/api";
 import { Plus, Calendar, Users } from "lucide-react";
 
 const bookingFormSchema = z.object({
