+++ client/src/lib/api.ts
@@ -1,0 +1,33 @@
+const API_BASE_URL = import.meta.env.VITE_API_URL || "";
+
+function withBase(url: string) {
+  return url.startsWith("http") ? url : `${API_BASE_URL}${url}`;
+}
+
+async function throwIfResNotOk(res: Response) {
+  if (!res.ok) {
+    const text = (await res.text()) || res.statusText;
+    throw new Error(`${res.status}: ${text}`);
+  }
+}
+
+export async function apiRequest(
+  method: string,
+  url: string,
+  data?: unknown,
+): Promise<Response> {
+  const res = await fetch(withBase(url), {
+    method,
+    headers: data ? { "Content-Type": "application/json" } : {},
+    body: data ? JSON.stringify(data) : undefined,
+    credentials: "include",
+  });
+
+  await throwIfResNotOk(res);
+  return res;
+}
+
+export function apiUrl(path: string) {
+  return withBase(path);
+}
+
