+++ client/src/pages/admin/ceo-dashboard.tsx
@@ -34,7 +34,7 @@ import {
 } from "lucide-react";
 import { useState } from "react";
 import { useMutation, useQueryClient } from "@tanstack/react-query";
-import { apiRequest } from "@/lib/queryClient";
+import { apiRequest } from "@/lib/api";
 import { useToast } from "@/hooks/use-toast";
 import SuperAdminRoute from "@/components/superadmin-route";
