+++ client/vite.config.ts
@@ -36,6 +36,9 @@ export default defineConfig(({ mode }) => {
       },
     },
     server: {
+      fs: {
+        allow: ['..'],
+      },
       port: 3000,
       proxy: {
         '/api': {
