+++ server/package.json
@@ -5,7 +5,7 @@
   "scripts": {
     "dev": "NODE_ENV=development tsx index.ts",
     "build": "tsc",
-    "start": "node dist/index.js",
+    "start": "node dist/server/index.js",
     "check": "tsc --noEmit"
   },
   "dependencies": {
