+++ server/routes.ts
@@ -6,7 +6,7 @@ import bcrypt from "bcrypt";
 import { storage } from "./storage";
 import { insertBookingSchema, insertReviewSchema, insertActivitySchema } from "@shared/schema";
 import { whatsappService } from "./whatsapp-service";
-import { z, ZodError } from "zod";
+import { z } from "zod";
 import {
   authRateLimit,
   adminApiRateLimit,
@@ -213,7 +213,7 @@ export async function registerRoutes(app: Express): Promise<Server> {
 
       res.status(201).json(booking);
     } catch (error: unknown) {
-      if (error instanceof ZodError) {
+      if (error instanceof z.ZodError) {
         return res.status(400).json({ message: "Validation error" });
       }
       console.error("Error creating booking:", error);
@@ -563,7 +563,7 @@ export async function registerRoutes(app: Express): Promise<Server> {
       
       res.status(201).json(activity);
     } catch (error: unknown) {
-      if (error instanceof ZodError) {
+      if (error instanceof z.ZodError) {
         return res.status(400).json({ message: "Validation error" });
       }
       console.error("Error creating activity:", error);
@@ -587,7 +587,7 @@ export async function registerRoutes(app: Express): Promise<Server> {
       
       res.json(activity);
     } catch (error: unknown) {
-      if (error instanceof ZodError) {
+      if (error instanceof z.ZodError) {
         return res.status(400).json({ message: "Validation error" });
       }
       console.error("Error updating activity:", error);
@@ -644,10 +644,10 @@ export async function registerRoutes(app: Express): Promise<Server> {
       const review = await storage.createReview(validatedData);
       res.status(201).json(review);
     } catch (error: unknown) {
-      if (error instanceof ZodError) {
+      if (error instanceof z.ZodError) {
         return res.status(400).json({ 
           message: "Validation error", 
-          errors: error.errors 
+          errors: error.issues
         });
       }
       console.error("Error creating review:", error);
